requirements.txt
Django==3.2.12
django-rest-framework==0.1.0
djangorestframework==3.14.0
packaging==23.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.6.2
python-dateutil==2.8.1
sqlparse==0.4.2
Django-extensions==3.2.3

